Fixing warnings

+++ GameEngine/Demos/SolarSystem/DemoScene.cpp
@@ -232,8 +232,11 @@ bool 			DemoScene::userStart()
 		{
 			planets[i] = createPlanet((std::rand() % 200) / 100.0f
 				, (std::rand() % 200) / 100.0f,
-				glm::vec3(std::rand() % 300 - 150, std::rand() % 300 - 150, std::rand() % 300 - 150),
-				glm::vec3(std::rand() % 10 + 10), "basic", "texture__SunTexture");
+				glm::vec3(static_cast<float>(std::rand() % 300 - 150),
+				static_cast<float>(std::rand() % 300 - 150),
+				static_cast<float>(std::rand() % 300 - 150)),
+				glm::vec3(static_cast<float>(std::rand() % 10 + 10)),
+				"basic", "texture__SunTexture");
 			if (i == 0)
 				sun->addChild(planets[i]);
 			else
